+++ ft_strlen.c
@@ -6,13 +6,15 @@
 /*   By: matoledo <matoledo@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/08 14:04:44 by matoledo          #+#    #+#             */
-/*   Updated: 2025/04/09 13:09:06 by matoledo         ###   ########.fr       */
+/*   Updated: 2025/04/09 20:33:53 by matoledo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-unsigned long	ft_strlen(const char *c)
+#include "libft.h"
+
+size_t	ft_strlen(const char *c)
 {
-	unsigned long	counter;
+	size_t	counter;
 
 	counter = 0;
 	while (*c++)
